refactor: 클래스명에 버저닝 적용
