+++ backend/src/main/java/com/festago/admin/dto/ArtistV1UpdateRequest.java
@@ -1,8 +1,8 @@
-package com.festago.artist.dto;
+package com.festago.admin.dto;
 
 import jakarta.validation.constraints.NotBlank;
 
-public record ArtistV1CreateRequest(
+public record ArtistV1UpdateRequest(
     @NotBlank(message = "아티스트 이름은 비어있을 수 없습니다.")
     String name,
     @NotBlank(message = "아티스트 이미지는 비어있을 수 없습니다.")
